+++ libs/src/ui/element/Textarea/Textarea.vue
@@ -74,20 +74,20 @@ const hintClass = computed(() => {
 </script>
 
 <template>
-	<div :class="{'textarea-container': true, [props.className]: !!props.className}">
+	<div :class="{'ded-textarea-container': true, [props.className]: !!props.className}">
 		<!-- 多行輸入框標題 -->
 		<template v-if="props.label">
-			<label :class="['textarea-label', {'textarea-disable': props.isDisabled}]" for="id">
+			<label :class="['ded-textarea-label', {'ded-textarea-disable': props.isDisabled}]" for="id">
 				{{ props.label }}
 			</label>
 		</template>
 
-		<div :class="['textarea-group',
-			{'textarea-disable': props.isDisabled},
-			( props.hint.error.length > 0 && `textarea-border-${hintClass}`)]">
+		<div :class="['ded-textarea-group',
+			{'ded-textarea-disable': props.isDisabled},
+			( props.hint.error.length > 0 && `ded-textarea-border-${hintClass}`)]">
 			<!-- 多行輸入框 -->
 			<textarea id="id"
-			          :class="['textarea', {'textarea-disable': props.isDisabled}]"
+			          :class="['ded-textarea', {'ded-textarea-disable': props.isDisabled}]"
 			          :maxlength="props.limit > 0 ? props.limit : undefined"
 			          :placeholder="props.placeholder"
 			          v-model="modelValue"
@@ -95,13 +95,13 @@ const hintClass = computed(() => {
 
 			<!-- 多行輸入字數提示 -->
 			<template v-if="modelValue.length > 0 && props.limit === 0">
-				<small :class="['textarea-hint-count', {'textarea-disable': props.isDisabled}]">
+				<small :class="['ded-textarea-hint-count', {'ded-textarea-disable': props.isDisabled}]">
 					{{ modelValue.length }}
 				</small>
 			</template>
 
 			<template v-else>
-				<small :class="['textarea-hint-count', {'textarea-disable': props.isDisabled}]">
+				<small :class="['ded-textarea-hint-count', {'ded-textarea-disable': props.isDisabled}]">
 					{{ modelValue.length > 0 ? `${modelValue.length} / ${props.limit}` : '' }}
 				</small>
 			</template>
@@ -109,7 +109,7 @@ const hintClass = computed(() => {
 
 		<!-- 多行輸入說明文字與提示 -->
 		<template v-if="props.hint.error.length > 0 || props.hint.description.length > 0">
-			<small :class="['textarea-hint', {'textarea-disable': props.isDisabled}, `textarea-hint-${hintClass}`]">
+			<small :class="['ded-textarea-hint', {'ded-textarea-disable': props.isDisabled}, `ded-textarea-hint-${hintClass}`]">
 				{{ props.hint.error.length > 0 ? props.hint.error : props.hint.description }}
 			</small>
 		</template>