+++ libs/src/ui/element/Tooltip/Tooltip.vue
@@ -46,8 +46,10 @@ const hideTooltip = () => {
 };
 
 // 功能 計算位置
-const updateTooltipPosition = () => {
+const updateTooltipPosition = async () => {
 	if (!tooltipTriggerRef.value || !tooltipContentRef.value || !visible.value) return;
+
+	await nextTick();
 
 	const triggerElement = tooltipTriggerRef.value.getBoundingClientRect();
 	const tooltipElement = tooltipContentRef.value.getBoundingClientRect();
@@ -57,58 +59,58 @@ const updateTooltipPosition = () => {
 	switch (props.placement) {
 		//-- [ top ] --//
 		case 'top-right':
-			top = (triggerElement.top - tooltipElement.height - 12);
+			top = (triggerElement.top - tooltipElement.height - 8);
 			left = triggerElement.right - tooltipElement.width;
 			break;
 		case 'top':
-			top = (triggerElement.top - tooltipElement.height - 12);
+			top = (triggerElement.top - tooltipElement.height - 8);
 			left = triggerElement.left + (triggerElement.width / 2) - (tooltipElement.width / 2);
 			break;
 		case 'top-left':
-			top = (triggerElement.top - tooltipElement.height - 12);
+			top = (triggerElement.top - tooltipElement.height - 8);
 			left = triggerElement.left;
 			break;
 
 		//-- [ right ] --//
 		case 'right-top':
 			top = triggerElement.top;
-			left = (triggerElement.right + 12);
+			left = (triggerElement.right + 8);
 			break;
 		case 'right':
 			top = triggerElement.top + (triggerElement.height / 2) - (tooltipElement.height / 2);
-			left = (triggerElement.right + 12);
+			left = (triggerElement.right + 8);
 			break;
 		case 'right-bottom':
 			top = (triggerElement.bottom - tooltipElement.height);
-			left = (triggerElement.right + 12);
+			left = (triggerElement.right + 8);
 			break;
 
 		//-- [ bottom ] --//
 		case 'bottom-right':
-			top = (triggerElement.bottom + 12);
+			top = (triggerElement.bottom + 8);
 			left = triggerElement.right - tooltipElement.width;
 			break;
 		case 'bottom':
-			top = (triggerElement.bottom + 12);
+			top = (triggerElement.bottom + 8);
 			left = triggerElement.left + (triggerElement.width / 2) - (tooltipElement.width / 2);
 			break;
 		case 'bottom-left':
-			top = (triggerElement.bottom + 12);
+			top = (triggerElement.bottom + 8);
 			left = triggerElement.left;
 			break;
 
 		//-- [ left ] --//
 		case 'left-top':
 			top = triggerElement.top;
-			left = (triggerElement.left - tooltipElement.width - 12);
+			left = (triggerElement.left - tooltipElement.width - 8);
 			break;
 		case 'left':
 			top = triggerElement.top + (triggerElement.height / 2) - (tooltipElement.height / 2);
-			left = (triggerElement.left - tooltipElement.width - 12);
+			left = (triggerElement.left - tooltipElement.width - 8);
 			break;
 		case 'left-bottom':
 			top = (triggerElement.bottom - tooltipElement.height);
-			left = (triggerElement.left - tooltipElement.width - 12);
+			left = (triggerElement.left - tooltipElement.width - 8);
 			break;
 	}
 
@@ -133,15 +135,13 @@ onBeforeUnmount(() => {
 <template>
 	<div
 		ref="tooltipTriggerRef"
-		:class="{'tooltip-trigger': true, [props.className]: !!props.className}"
+		:class="{'ded-tooltip-container': true, [props.className]: !!props.className}"
 		@mouseenter="showTooltip"
 		@mouseleave="hideTooltip"
 		:aria-describedby="tooltipId"
 	>
-		<!-- Tooltip - slot 包裝 -->
 		<slot></slot>
 	</div>
-
 	<Teleport to="body">
 		<Transition name="fade">
 			<div
@@ -149,12 +149,17 @@ onBeforeUnmount(() => {
 				:id="tooltipId"
 				ref="tooltipContentRef"
 				:style="tooltipStyles"
-				class="tooltip-content"
-				:class="props.placement"
+				class="ded-tooltip ded-tooltip-content"
+				:class="`ded-tooltip-${props.placement}`"
 			>
-				{{props.content}}
-				<!-- Tooltip - 箭頭 -->
-				<div v-if="props.showArrow" :class="`tooltip-arrow ${props.placement}`"></div>
+				{{ props.content }}
+				<div
+					v-if="props.showArrow"
+					class="ded-tooltip-arrow"
+					:class="`ded-tooltip-arrow-${props.placement}`"
+				>
+					<div class="ded-tooltip-arrow-shape"></div>
+				</div>
 			</div>
 		</Transition>
 	</Teleport>